sayfalar tamamlandı

+++ AenHospital/AenHospital/Services/Patient/Interface/IPatientOrderService.cs
@@ -1,6 +1,7 @@
 ﻿using AenHospital.Models;
 using System;
 using System.Collections.Generic;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -9,6 +10,6 @@ namespace AenHospital.Services.Patient.Interface
     public interface IPatientOrderService
     {
         // Hasta istekleri listesi
-        Task<List<PatientOrder>> GetPatientOrderListAsync(int ptn);
+        Task<List<PatientOrder>> GetPatientOrderListAsync(BigInteger ptn);
     }
 }
